Moved query editor related directives in the core package
(cherry picked from commit 79745f8)

Fix paginations module name
