+++ src/js/angular/core/directives/queryeditor/templates/query-editor.html
