+++ src/js/angular/core/directives/queryeditor/templates/query-sample.html
