+++ src/js/angular/export/app.js
@@ -1,7 +1,7 @@
 import 'angular/core/services';
 import 'angular/core/directives';
 import 'angular/export/controllers';
-import 'angular/directives/paginations';
+import 'angular/core/directives/paginations';
 import 'angular/core/services/repositories.service';
 import 'lib/FileSaver-patch';
 
@@ -11,7 +11,7 @@ const modules = [
     'ngRoute',
     'ngCookies',
     'graphdb.framework.impex.export.controllers',
-    'graphdb.framework.directives.paginations',
+    'graphdb.framework.core.directives.paginations',
     'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives'
 ];
