+++ src/js/angular/namespaces/app.js
@@ -3,7 +3,7 @@ import 'angular/core/directives';
 import 'angular/core/services/repositories.service';
 import 'angular/namespaces/controllers';
 import 'angular/namespaces/directives';
-import 'angular/directives/paginations';
+import 'angular/core/directives/paginations';
 import 'angular-xeditable/dist/js/xeditable.min';
 
 const modules = [
@@ -16,7 +16,7 @@ const modules = [
     'graphdb.framework.namespaces.directives',
     'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives',
-    'graphdb.framework.directives.paginations'
+    'graphdb.framework.core.directives.paginations'
 ];
 
 const namespacesApp = angular.module('graphdb.framework.namespaces', modules);
